+++ src/SongManager.hpp
@@ -1,5 +1,8 @@
 #pragma once
 
+#define ECLIPSE_MODULES_HPP
+#include <eclipse.eclipse-menu/include/eclipse.hpp>
+
 enum class SongType {
 	Regular,
 	Favorited,
@@ -124,6 +127,9 @@ class SongManager {
 	void setOsu(const bool value);
 	bool getOsu() const;
 	unsigned long getHashedCurrentSong() const;
+
+	std::optional<eclipse::components::Label> eclipseSongNameLabel;
+	std::optional<eclipse::components::Label> eclipseSongDurationLabel;
 
   private:
 	SongManager();